+++ src/backend/user-service/src/main/java/com/capstone/userservice/domain/user/dto/UserDeleteRequest.java
@@ -1,0 +1,14 @@
+package com.capstone.userservice.domain.user.dto;
+
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@AllArgsConstructor
+public class UserDeleteRequest {
+    private String email;
+    private String password;
+}
